requirements.txt
PyPDF2>=3.0.0
reportlab>=4.0.0
Pillow>=10.0.0